[all] avoid using atomic_flag and fix struct alignments for ABI reasons
We need to be compatible with hosts/clients compiled using MSVC with the
incoming Idd driver. As such the packing of structs and size of members
must be consistent between compilers. Ensuring 64-bit alignment of
members and using a uint32_t as the lock is required as there is no
available equivalent of atomic_flag in MSVC 2019 that we can make use of