[all] do not specify implicit alignment for MSVC
MSVC pack and declspec(align()) do not work in the same way as gcc's
align attribute. Each member of the struct needs to be individually
aligned in MSVC and doing so results in a struct that is incompatible
with gcc's. For now we will just allow it to use the default as testing
shows it's compatible.